Calling FRaC now, Tried adding parsing

+++ csax/csax_cpp/csaxfuncs.h
@@ -15,4 +15,4 @@ struct SampleList {
 };
 
 void runCSAX(SampleList traindata, SampleList testdata, string testfile,
-        string genesets_file, int num_bags, float gamma);
+        string genesets_file, int num_bags, double gamma);
